backend/sockets/pyproject.toml
[project]
name = "sockets"
version = "0.1.0"
description = ""
authors = [
    {name = "ant100pshnik",email = "mypochtav@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "redis (>=7.1.0,<8.0.0)",
    "faststream[cli,rabbit] (>=0.6.3,<0.7.0)",
    "dishka (>=1.7.2,<2.0.0)",
    "pymongo (>=4.15.4,<5.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "python-socketio[asyncio] (>=5.15.0,<6.0.0)",
    "boto3 (>=1.41.2,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "aiohttp (>=3.13.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
